+++ kivakit-snippets/src/main/java/com/telenav/kivakit/snippets/Snippet14.java
@@ -1,7 +1,6 @@
 package com.telenav.kivakit.snippets;
 
-import com.telenav.kivakit.conversion.core.time.HumanizedLocalDateTimeConverter;
-import com.telenav.kivakit.core.registry.RegistryTrait;
+import com.telenav.kivakit.conversion.core.time.kivakit.HumanizedKivaKitLocalDateTimeConverter;
 
 import static com.telenav.kivakit.core.time.LocalTime.now;
 
@@ -15,7 +14,7 @@ public class Snippet14 extends BaseSnippet
     @Override
     protected Object onSnippet()
     {
-        var converter = new HumanizedLocalDateTimeConverter(this);
+        var converter = new HumanizedKivaKitLocalDateTimeConverter(this);
         return converter.unconvert(now().asLocalTime())
             + ", "
             + converter.convert("Today 6.00pm");
