+++ kivakit-snippets/src/main/java/com/telenav/kivakit/snippets/Snippet16.java
@@ -1,13 +1,10 @@
 package com.telenav.kivakit.snippets;
 
 import com.telenav.kivakit.application.Application;
-import com.telenav.kivakit.core.object.Lazy;
 import com.telenav.kivakit.core.os.ConsoleTrait;
-import com.telenav.kivakit.serialization.gson.factory.KivaKitCoreGsonFactory;
 
 import static com.telenav.kivakit.core.collections.list.StringList.split;
-import static com.telenav.kivakit.core.object.Lazy.lazy;
-import static com.telenav.kivakit.network.http.secure.SecureHttpNetworkLocation.secureHttpNetworkLocation;
+import static com.telenav.kivakit.network.https.HttpsNetworkLocation.secureHttpNetworkLocation;
 
 public class Snippet16 extends Application implements ConsoleTrait
 {
