+++ kivakit-snippets/src/main/java/com/telenav/kivakit/snippets/Snippet8.java
@@ -8,7 +8,7 @@ import static com.telenav.kivakit.core.collections.list.StringList.stringList;
 import static com.telenav.kivakit.core.string.AsciiArt.line;
 import static com.telenav.kivakit.filesystem.File.file;
 import static com.telenav.kivakit.network.core.Host.host;
-import static com.telenav.kivakit.network.http.secure.SecureHttpNetworkLocation.secureHttpNetworkLocation;
+import static com.telenav.kivakit.network.https.HttpsNetworkLocation.secureHttpNetworkLocation;
 
 public class Snippet8 extends BaseSnippet implements
     TryTrait,
